+++ powerauth-restful-security-spring/src/main/java/com/wultra/security/powerauth/rest/api/spring/controller/v3/ActivationController.java
@@ -184,13 +184,13 @@ public class ActivationController {
     }
 
     /**
-     * Remove activation.
-     * @param request Remove activation request.
+     * Rename activation.
+     * @param request Rename activation request.
      * @param auth PowerAuth authentication.
      * @return Activation detail response.
      * @throws PowerAuthCodeInvalidException In case the authentication code validation fails.
      * @throws PowerAuthInvalidRequestException In case request is invalid.
-     * @throws PowerAuthActivationException In case retrieving activation detail fails.
+     * @throws PowerAuthActivationException In case renaming activation fails.
      */
     @PostMapping("rename")
     @PowerAuth(resourceId = "/pa/activation/rename", authenticationCodeType = {